chore(deps): rusqlite 0.34→0.39 + workspace dep audit
## Why

Witchcraft pins `rusqlite ^0.39` (libsqlite3-sys ^0.37); this workspace
pinned `rusqlite 0.34`. libsqlite3-sys's `links = "sqlite3"` allows
exactly one version per dependency graph, so the optional Witchcraft
git dep alone made `cargo check -p leyline-cli` fail in any workspace
that wanted to add it. This PR bumps rusqlite to 0.39 and folds in a
broader audit of stale pins surfaced by `cargo outdated`.

## Bumped (explicit Cargo.toml edits)

| Crate       | From   | To       | Notes                                                      |
|-------------|--------|----------|------------------------------------------------------------|
| rusqlite    | 0.34   | 0.39     | Witchcraft unblock + `DatabaseName::Main` → `"main"` (new `Name` trait, blanket `&str` impl); `usize`/`u64` bind through `i64` (those `ToSql`/`FromSql` impls were dropped) |
| capnp       | =0.25.0| =0.25.4  | Patch; ADR-0014 §3 exact-pin convention preserved across cli-lib / lsp / public-schema / schema-capnp |
| dirs        | 5      | 6        | Major; surface used is `home_dir()` only (stable across 5→6)|
| which       | 7      | 8        | Major; surface used is `which()`/`Which::new()` (stable)    |
| criterion   | 0.5    | 0.8      | Dev-only; benches compile unchanged                         |
| lsp-types   | 0.95   | 0.97     | `Url` renamed to `Uri` (newtype around `fluent_uri::Uri<String>`); `Url::parse(s).unwrap()` → `Uri::from_str(s).unwrap()` |

## Picked up via `cargo update` (wildcard pins, no Cargo.toml edits)

tokio 1.50 → 1.52 · clap 4.6.0 → 4.6.1 · libc 0.2.183 → 0.2.186 ·
sqlite-vec 0.1.7 → 0.1.9 · tree-sitter 0.26.7 → 0.26.8

## Deferred (separate PR or skip — each warrants its own migration
## and test review)

thiserror 1→2 (ecosystem-wide derive change), nalgebra 0.32→0.34
(coupled to sheaf math — DVector/DMatrix surface), sha2 0.10→0.11
(crypto stack), der/spki/const-oid (signing pipeline), jj-lib
0.38→0.41 (VCS adapter), fuser 0.15→0.17 (FUSE op signatures),
nfsserve 0.10→0.11, uv-* 0.0.29→0.0.47 (pyproject parsing, 18-version
jump), toml 0.8→1.0.

## Verification

- `task ci` clean (30+ test groups all `ok`, including the new
  blackbox tests landed in #32)
- Sheaf event-push regression test (`event_push_blackbox_test.rs`)
  still green post-bump — confirms the v0.4.3 fix isn't disturbed

+++ rs/ll-open/cli-lib/tests/cold_parse_perf_regression.rs
@@ -108,10 +108,12 @@ fn build_corpus(target: &Path, copies: usize) -> std::io::Result<PathBuf> {
 /// node-emission strategy evolves (e.g. if we ever start storing
 /// additional per-token rows).
 fn count_rows(db_path: &Path) -> rusqlite::Result<u64> {
+    // rusqlite 0.39 dropped `FromSql for u64`; read via `i64` then cast.
+    // `COUNT(*)` is non-negative so the cast is total.
     let conn = rusqlite::Connection::open(db_path)?;
-    let nodes: u64 = conn.query_row("SELECT COUNT(*) FROM nodes", [], |r| r.get(0))?;
-    let ast: u64 = conn.query_row("SELECT COUNT(*) FROM _ast", [], |r| r.get(0))?;
-    Ok(nodes + ast)
+    let nodes: i64 = conn.query_row("SELECT COUNT(*) FROM nodes", [], |r| r.get(0))?;
+    let ast: i64 = conn.query_row("SELECT COUNT(*) FROM _ast", [], |r| r.get(0))?;
+    Ok((nodes + ast) as u64)
 }
 
 #[tokio::test]
